Rename everything to match "transfer referral" name
