+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/jobs/oneoff/transferreferrals/TransferReferralsJobConfiguration.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jobs.oneoff.movereferrals
+package uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jobs.oneoff.transferreferrals
 
 import org.hibernate.SessionFactory
 import org.springframework.batch.core.Job
@@ -8,34 +8,40 @@ import org.springframework.batch.core.configuration.annotation.JobBuilderFactory
 import org.springframework.batch.core.configuration.annotation.JobScope
 import org.springframework.batch.core.configuration.annotation.StepBuilderFactory
 import org.springframework.batch.core.job.DefaultJobParametersValidator
-import org.springframework.batch.item.ItemWriter
 import org.springframework.batch.item.database.HibernateCursorItemReader
 import org.springframework.batch.item.database.builder.HibernateCursorItemReaderBuilder
 import org.springframework.beans.factory.annotation.Value
+import org.springframework.boot.ApplicationRunner
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.springframework.stereotype.Component
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jobs.oneoff.OnStartupJobLauncherFactory
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jpa.entity.Referral
 
 @Configuration
 @EnableBatchProcessing
-class MoveReferralsJobConfiguration(
+class TransferReferralsJobConfiguration(
   private val jobBuilderFactory: JobBuilderFactory,
   private val stepBuilderFactory: StepBuilderFactory,
-  private val listener: MoveReferralsJobListener,
+  private val listener: TransferReferralsJobListener,
+  private val onStartupJobLauncherFactory: OnStartupJobLauncherFactory,
 ) {
   private val transferSignalSubject = "Authority to transfer case"
   private val transferSignalText = "On behalf of Advance Charity Ltd, I request and give permission for this referral to be transferred to Women in Prison, as of December 2022."
 
   @Bean
+  fun transferReferralsJobLauncher(transferReferralsJob: Job): ApplicationRunner {
+    return onStartupJobLauncherFactory.makeBatchLauncher(transferReferralsJob)
+  }
+
+  @Bean
   @JobScope
-  fun moveReferralsReader(
+  fun transferReferralsReader(
     @Value("#{jobParameters['fromContract']}") fromContract: String,
     @Value("#{jobParameters['toContract']}") toContract: String,
     sessionFactory: SessionFactory,
   ): HibernateCursorItemReader<Referral> {
     return HibernateCursorItemReaderBuilder<Referral>()
-      .name("moveReferralsReader")
+      .name("transferReferralsReader")
       .sessionFactory(sessionFactory)
       .queryString(
         "SELECT r FROM Referral r JOIN CaseNote n ON n.referralId = r.id " +
@@ -53,7 +59,7 @@ class MoveReferralsJobConfiguration(
   }
 
   @Bean
-  fun moveReferralsJob(updateInterventionIdStep: Step): Job {
+  fun transferReferralsJob(transferReferralToInterventionStep: Step): Job {
     val validator = DefaultJobParametersValidator()
     validator.setRequiredKeys(
       arrayOf(
@@ -63,31 +69,24 @@ class MoveReferralsJobConfiguration(
       )
     )
 
-    return jobBuilderFactory["moveReferralsJob"]
+    return jobBuilderFactory["transferReferralsJob"]
       .validator(validator)
       .listener(listener)
-      .start(updateInterventionIdStep)
+      .start(transferReferralToInterventionStep)
       .build()
   }
 
   @Bean
-  fun updateInterventionIdStep(
-    moveReferralsReader: HibernateCursorItemReader<Referral>,
-    processor: MoveReferralsProcessor,
-    writer: ItemWriter<Referral?>,
+  fun transferReferralToInterventionStep(
+    transferReferralsReader: HibernateCursorItemReader<Referral>,
+    processor: TransferReferralsProcessor,
+    writer: TransferReferralsWriter,
   ): Step {
-    return stepBuilderFactory.get("updateInterventionIdStep")
+    return stepBuilderFactory.get("transferReferralToInterventionStep")
       .chunk<Referral, Referral>(10)
-      .reader(moveReferralsReader)
+      .reader(transferReferralsReader)
       .processor(processor)
       .writer(writer)
       .build()
-  }
-}
-
-@Component
-class noOpWriter() : ItemWriter<Referral> {
-  override fun write(items: MutableList<out Referral>) {
-    /* no-op */
   }
 }
