+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/jobs/oneoff/transferreferrals/TransferReferralsJobListener.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jobs.oneoff.movereferrals
+package uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jobs.oneoff.transferreferrals
 
 import mu.KLogging
 import net.logstash.logback.argument.StructuredArguments.kv
@@ -8,7 +8,7 @@ import org.springframework.batch.core.JobExecutionListener
 import org.springframework.stereotype.Component
 
 @Component
-class MoveReferralsJobListener : JobExecutionListener {
+class TransferReferralsJobListener : JobExecutionListener {
   companion object : KLogging()
 
   override fun beforeJob(jobExecution: JobExecution) {
