+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/jobs/oneoff/transferreferrals/TransferReferralsWriter.kt
@@ -1,0 +1,11 @@
+package uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jobs.oneoff.transferreferrals
+
+import org.springframework.batch.item.ItemWriter
+import org.springframework.stereotype.Component
+import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jpa.entity.Referral
+
+@Component
+class TransferReferralsWriter : ItemWriter<Referral?> {
+  override fun write(items: MutableList<out Referral?>) {
+  }
+}
